+++ tests/Command/AuditCommandTest.php
@@ -89,6 +89,10 @@ final class AuditCommandTest extends TestCase
         );
 
         foreach ($files as $file) {
+            if (! $file instanceof \SplFileInfo) {
+                continue;
+            }
+
             $file->isDir() ? rmdir($file->getPathname()) : unlink($file->getPathname());
         }
